refactor: remove e2ee-server app and restructure E2EE components
- Remove standalone e2ee-server app from apps/e2ee-server/
- Move E2EE bridge components to packages/kit-bg/src/services/ServicePrimeTransfer/e2ee/
- Reorganize E2EE client-to-client and server API components
- Update import paths in ServicePrimeTransfer to use new e2ee/ subdirectory
- Convert RoomManager class to IRoomManager interface definition
- Clean up unused dependencies and configurations
