+++ .eslintrc.js
@@ -85,11 +85,6 @@ const restrictedImportsPatterns = [
     group: ['**/v4localDBStoreNames.native'],
     message: 'import v4localDBStoreNames instead ',
   },
-  {
-    group: ['**/e2ee-server/**', '@onekeyhq/e2ee-server/**'],
-    message:
-      'Import from e2ee-server is not allowed because e2ee-server is not client-side code',
-  },
   //
 ];
 const tsRules = {
