+++ .gitignore
@@ -8,7 +8,6 @@ __generated__
 .vscode/
 .chat/
 .claude/
-apps/e2ee-server/dist/
 
 dist
 build-electron
