--- apps/e2ee-server/.yarnrc.yml
@@ -1,1 +1,0 @@
-nodeLinker: node-modules
