--- apps/e2ee-server/env.example
@@ -1,16 +1,0 @@
-# E2EE Server Configuration
-
-# Server port (default: 3868)
-PORT=3868
-
-# CORS allowed origins, comma-separated
-CORS_ORIGINS=http://localhost:3000,http://localhost:3001,http://localhost:3868
-
-# Maximum users per room (default: 2)
-MAX_USERS_PER_ROOM=2
-
-# Room timeout in milliseconds - default 1 hour
-ROOM_TIMEOUT=3600000
-
-# Maximum message size in bytes - default 1MB
-MAX_MESSAGE_SIZE=1048576
