--- apps/e2ee-server/nodemon.json
@@ -1,8 +1,0 @@
-{
-  "watch": ["src"],
-  "ext": "ts,js,json",
-  "exec": "ts-node src/server.ts",
-  "env": {
-    "NODE_ENV": "development"
-  }
-}
