+++ packages/kit-bg/src/services/ServicePrimeTransfer/e2ee/JsBridgeE2EEClientToClient.ts
@@ -1,5 +1,4 @@
 import { JsBridgeBase } from '@onekeyfe/cross-inpage-provider-core';
-import { IJsBridgeMessageTypes } from '@onekeyfe/cross-inpage-provider-types';
 
 import type {
   IJsBridgeConfig,
