+++ packages/kit-bg/src/services/ServicePrimeTransfer/e2ee/e2eeClientToClientApiProxy.ts
@@ -1,6 +1,6 @@
 /* eslint-disable no-restricted-syntax */
 
-import { RemoteApiProxyBase } from '../../apis/RemoteApiProxyBase';
+import { RemoteApiProxyBase } from '../../../apis/RemoteApiProxyBase';
 
 import { JsBridgeE2EEClientToClient } from './JsBridgeE2EEClientToClient';
